+++ src/components/form/JSONSchemaForm.tsx
@@ -3,12 +3,12 @@ import { Fields } from '@remoteoss/json-schema-form-old';
 import React, { Fragment } from 'react';
 
 import { fieldsMap } from '@/src/components/form/fields/fieldsMapping';
-import { SupportedTypes } from '@/src/components/form/fields/types';
 import { Statement } from '@/src/components/form/Statement';
 import { ForcedValueField } from '@/src/components/form/fields/ForcedValueField';
 import { Components, JSFFieldset } from '@/src/types/remoteFlows';
 import { getFieldsWithFlatFieldsets } from './utils';
 import { StatementComponentProps } from '@/src/types/fields';
+import { useFormContext } from 'react-hook-form';
 
 type JSONSchemaFormFieldsProps = {
   fields: Fields;
@@ -33,6 +33,8 @@ export const JSONSchemaFormFields = ({
   fieldValues,
   components,
 }: JSONSchemaFormFieldsProps) => {
+  const { setValue } = useFormContext();
+
   if (!fields || fields.length === 0) return null;
 
   const maybeFieldWithFlatFieldsets =
@@ -76,10 +78,17 @@ export const JSONSchemaFormFields = ({
           const { Component } = field as {
             Component: React.ComponentType<any>;
           };
-          return <Component key={field.name as string} {...field} />;
+          return (
+            <Component
+              key={field.name as string}
+              setValue={(value: unknown) => setValue(field.name, value)}
+              {...field}
+            />
+          );
         }
 
-        let FieldComponent = fieldsMap[field.inputType as SupportedTypes];
+        let FieldComponent =
+          fieldsMap[field.inputType as keyof typeof fieldsMap];
 
         if (!FieldComponent) {
           return (
@@ -120,7 +129,7 @@ export const JSONSchemaFormFields = ({
             <FieldComponent
               {...field}
               component={
-                components && components[field.inputType as SupportedTypes]
+                components && components[field.inputType as keyof Components]
               }
             />
             {field.statement ? (