+++ src/components/form/fields/FieldSetField.tsx
@@ -2,16 +2,12 @@ import { useFormContext } from 'react-hook-form';
 import { Fragment, useEffect, useRef } from 'react';
 import { baseFields } from '@/src/components/form/fields/baseFields';
 import { cn, sanitizeHtml } from '@/src/lib/utils';
-import {
-  $TSFixMe,
-  Components,
-  FieldSetToggleComponentProps,
-} from '@/src/types/remoteFlows';
+import { $TSFixMe, Components } from '@/src/types/remoteFlows';
 import { Statement } from '@/src/components/form/Statement';
 import { useFormFields } from '@/src/context';
-import { Button } from '@/src/components/ui/button';
 import { ZendeskTriggerButton } from '@/src/components/shared/zendesk-drawer/ZendeskTriggerButton';
-import { SupportedTypes } from './types';
+import { FieldsetToggleButtonDefault } from '@/src/components/form/fields/default/FieldsetToggleButtonDefault';
+import { BaseTypes, SupportedTypes } from './types';
 import { StatementComponentProps } from '@/src/types/fields';
 
 type FieldBase = {
@@ -68,26 +64,6 @@ export type FieldSetProps = {
     };
   } & Record<string, $TSFixMe>;
 };
-
-const DefaultToggleButton = ({
-  isExpanded,
-  onToggle,
-  className,
-  ...props
-}: FieldSetToggleComponentProps) => (
-  <Button
-    type='button'
-    className={cn(
-      'RemoteFlows__Button RemoteFlows__FieldSetField__ToggleButton',
-      className,
-    )}
-    variant='default'
-    onClick={onToggle}
-    {...props}
-  >
-    {isExpanded ? 'Remove' : 'Define'}
-  </Button>
-);
 
 export function FieldSetField({
   label,
@@ -162,7 +138,8 @@ export function FieldSetField({
     };
   }, [watchedValues, trigger, formState.isSubmitted, formState.submitCount]);
 
-  const ToggleComponent = formComponents?.fieldsetToggle || DefaultToggleButton;
+  const ToggleComponent =
+    formComponents?.fieldsetToggle || FieldsetToggleButtonDefault;
   const contentId = `${name}-content`;
   const headerId = `${name}-header`;
 
@@ -274,19 +251,24 @@ export function FieldSetField({
               }
               // We need to do the check after checking field.inputType === 'fieldset' or field.inputType === 'fieldset-flat'
               // circular dependency most likely
-              let FieldComponent =
-                baseFields[
-                  field.inputType as Exclude<
-                    SupportedTypes,
-                    'fieldset' | 'fieldset-flat'
-                  >
-                ];
+              let FieldComponent = baseFields[field.inputType as BaseTypes];
 
               if (field.Component) {
                 const { Component } = field as {
                   Component: React.ComponentType<$TSFixMe>;
                 };
-                return <Component key={field.name} {...field} />;
+                return (
+                  <Component
+                    key={field.name}
+                    setValue={(value: unknown) =>
+                      setValue(
+                        isFlatFieldset ? field.name : `${name}.${field.name}`,
+                        value,
+                      )
+                    }
+                    {...field}
+                  />
+                );
               }
 
               if (!FieldComponent) {
@@ -308,7 +290,9 @@ export function FieldSetField({
                   <FieldComponent
                     {...field}
                     name={`${isFlatFieldset ? field.name : `${name}.${field.name}`}`}
-                    component={components?.[field.inputType as SupportedTypes]}
+                    component={
+                      components?.[field.inputType as keyof Components]
+                    }
                   />
                   {field.extra ? field.extra : null}
                 </Fragment>