+++ src/default-components.ts
@@ -1,5 +1,5 @@
 import { Components } from '@/src/types/remoteFlows';
-import { FileUploadFieldDefault } from './FileUploadFieldDefault';
+import { FileUploadFieldDefault } from '@/src/components/form/fields/default/FileUploadFieldDefault';
 import { TextAreaFieldDefault } from '@/src/components/form/fields/default/TextAreaFieldDefault';
 import { SelectFieldDefault } from '@/src/components/form/fields/default/SelectFieldDefault';
 import { CountryFieldDefault } from '@/src/components/form/fields/default/CountryFieldDefault';
@@ -10,17 +10,29 @@ import { EmailFieldDefault } from '@/src/components/form/fields/default/EmailFie
 import { RadioGroupFieldDefault } from '@/src/components/form/fields/default/RadioGroupFieldDefault';
 import { MultiSelectFieldDefault } from '@/src/components/form/fields/default/MultiSelectFieldDefault';
 import { TextFieldDefault } from '@/src/components/form/fields/default/TextFieldDefault';
-
+import { FieldsetToggleButtonDefault } from '@/src/components/form/fields/default/FieldsetToggleButtonDefault';
+import { DrawerDefault } from '@/src/components/shared/drawer/DrawerDefault';
+import { ZendeskDrawerDefault } from '@/src/components/shared/zendesk-drawer/ZendeskDrawerDefault';
+import { TableFieldDefault } from '@/src/components/shared/table/TableFieldDefault';
+import { CheckboxFieldDefault } from '@/src/components/form/fields/default/CheckboxFieldDefault';
+import { WorkScheduleFieldDefault } from '@/src/components/form/fields/default/WorkScheduleFieldDefault';
+import { ButtonDefault } from '@/src/components/form/fields/default/ButtonDefault';
 /**
- * Default field components provided by RemoteFlows.
- * These components provide the standard UI implementations for all field types.
- * You can use these as-is or override specific fields with custom components.
+ * Default field components for RemoteFlows.
+ *
+ * This module acts as a reference for our users to know what are the components that are really overrided
+ *
  */
 export const defaultComponents: Components = {
+  button: ButtonDefault,
+  checkbox: CheckboxFieldDefault,
   countries: CountryFieldDefault,
   date: DatePickerFieldDefault,
+  drawer: DrawerDefault,
   email: EmailFieldDefault,
+  fieldsetToggle: FieldsetToggleButtonDefault,
   file: FileUploadFieldDefault,
+  // TODO: we have doubts multi-select, it's actually used
   'multi-select': MultiSelectFieldDefault,
   number: NumberFieldDefault,
   radio: RadioGroupFieldDefault,
@@ -28,4 +40,7 @@ export const defaultComponents: Components = {
   statement: StatementDefault,
   textarea: TextAreaFieldDefault,
   text: TextFieldDefault,
+  zendeskDrawer: ZendeskDrawerDefault,
+  table: TableFieldDefault,
+  'work-schedule': WorkScheduleFieldDefault,
 };