+++ src/flows/CostCalculator/CostCalculatorResetButton.tsx
@@ -1,8 +1,8 @@
-import { Button } from '@/src/components/ui/button';
 import { ButtonHTMLAttributes, PropsWithChildren } from 'react';
 import { useCostCalculatorContext } from './context';
 import { cn } from '@/src/lib/utils';
 import { useFormFields } from '@/src/context';
+import { ButtonDefault } from '@/src/components/form/fields/default/ButtonDefault';
 
 export function CostCalculatorResetButton({
   children,
@@ -12,7 +12,7 @@ export function CostCalculatorResetButton({
   const { form, formId, costCalculatorBag } = useCostCalculatorContext();
   const { components } = useFormFields();
 
-  const CustomButton = components?.button || Button;
+  const CustomButton = components?.button || ButtonDefault;
   return (
     <CustomButton
       {...props}