+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # @remoteoss/remote-flows
 
+## 1.3.0
+
+### Minor Changes
+
+- simplify currency conversion to have same data that estimation (#662) [#662](https://github.com/remoteoss/remote-flows/pull/662)
+- add new style export (#651) [#651](https://github.com/remoteoss/remote-flows/pull/651)
+- expose setValue prop to custom JSF components (#655) [#655](https://github.com/remoteoss/remote-flows/pull/655)
+- fix Components type (#660) [#660](https://github.com/remoteoss/remote-flows/pull/660)
+- add provisional start date inside the contractor onboarding flow (#663) [#663](https://github.com/remoteoss/remote-flows/pull/663)
+- add warning in contract details when dates are different (#664) [#664](https://github.com/remoteoss/remote-flows/pull/664)
+
+## 1.2.2
+
+### Patch Changes
+
+- add manual release triggering and simplify version detection (#642) [#642](https://github.com/remoteoss/remote-flows/pull/642)
+- extract checkbox and work schedule field defaults into separate components (#626) [#626](https://github.com/remoteoss/remote-flows/pull/626)
+- replace wildcard exports with explicit named exports (#652) [#652](https://github.com/remoteoss/remote-flows/pull/652)
+- consolidate helpCenter meta type into FieldDataProps (#650) [#650](https://github.com/remoteoss/remote-flows/pull/650)
+
+## 1.2.1
+
+### Patch Changes
+
+- extract default component implementations (#624) [#624](https://github.com/remoteoss/remote-flows/pull/624)
+- avoid sending not a number (#641) [#641](https://github.com/remoteoss/remote-flows/pull/641)
+- make sure updateEmployment follows the same rules than the GET (#646) [#646](https://github.com/remoteoss/remote-flows/pull/646)
+- fix json_schema_form params in updateEmployment (#644) [#644](https://github.com/remoteoss/remote-flows/pull/644)
+
 ## 1.2.0
 
 ### Minor Changes
@@ -99,6 +128,12 @@
 - Override json_schema_version for DEU contract details (#564) [#564](https://github.com/remoteoss/remote-flows/pull/564)
 - fix failed termination tests (#566) [#566](https://github.com/remoteoss/remote-flows/pull/566)
 - improve UX copy and date formatting (#565) [#565](https://github.com/remoteoss/remote-flows/pull/565)
+
+## 0.28.2
+
+### Fixes
+
+- fix updateEmployment mutation, passing incorrect param [0.28.2](https://github.com/remoteoss/remote-flows/releases/tag/v0.28.2)
 
 ## 0.28.1